requirements.txt
pandas
openpyxl
mysql-connector-python
python-dotenv